+++ src/Viaggi.h
@@ -22,7 +22,7 @@ typedef struct graphViaggi {
 
 GraphViaggi* AllocaGrafo();
 void InserisciVertice(GraphViaggi* G, char *citta);
-EdgeViaggi* CreaArco (char* citta, int prezzoAereo, int prezzoTreno, int tempoAereo, int  tempoTreno);
+EdgeViaggi* CreaArco (char* citta, float prezzoAereo, float prezzoTreno, int tempoAereo, int  tempoTreno);
 void addArco(GraphViaggi* grafo, int posizione, char *citta, float prezzoAereo, float prezzoTreno, int tempoAereo, int  tempoTreno);
 GraphViaggi* leggiFileViaggi(GraphViaggi* grafo);
 int isEmpty(GraphViaggi* grafo);
